+++ Frameworks/AIUtilities/Source/MVMenuButton.m
@@ -61,7 +61,7 @@
 // Control Size (Allows us to dynamically size for a small or big toolbar)
 - (void)setControlSize:(NSControlSize)inControlSize
 {
-	NSSize targetSize = NSMakeSize(0, 0);
+	NSSize targetSize;
 	NSSize bigImageSize;
 
 	controlSize = inControlSize;